+++ .scrutinizer.yml
@@ -66,4 +66,4 @@ checks:
     argument_type_checks: true
 
 filter:
-  paths: [code/*, tests/*]
+  paths: [src/*, tests/*]
