+++ .upgrade.yml
@@ -1,0 +1,2 @@
+mappings:
+  DependentDropdownField: Sheadawson\DependentDropdown\Forms\DependentDropdownField
