+++ _config.php
@@ -1,3 +1,1 @@
 <?php
-
-if(!defined('DEPENDENTDROPDOWNFIELD')) define('DEPENDENTDROPDOWNFIELD', rtrim(basename(dirname(__FILE__))));
